+++ src/cpp/traction_separation.cpp
@@ -13,7 +13,7 @@
 namespace tractionSeparation{
 
     errorOut computeCurrentDistance( const floatVector &Xi_1, const floatVector &Xi_2, const floatVector &D,
-                                     const floatVector &F,    const floatVector &chi,  const floatVector &grad_chi,
+                                     const floatVector &F,    const floatVector &chi,  const floatVector &gradChi,
                                      floatVector &d ){
         /*!
          * Compute the distance in the current configuration where
@@ -33,7 +33,7 @@ namespace tractionSeparation{
          * \param &D: The initial separation between the particles
          * \param &F: The deformation gradient \f$\frac{dx_i}{dX_I}\f$
          * \param &chi: The micro-deformation
-         * \param &grad_chi: the gradient of the micro deformation w.r.t. the reference spatial variable
+         * \param &gradChi: the gradient of the micro deformation w.r.t. the reference spatial variable
          * \param &d: The current separation between the particles
          */
 
@@ -50,7 +50,7 @@ namespace tractionSeparation{
 
                 for ( unsigned int J = 0; J < dX.size( ); J++ ){
 
-                    chi_2[ dX.size( ) * i + I ] += grad_chi[ dX.size( ) * dX.size( ) * i + dX.size( ) * I + J ] * dX[ J ];
+                    chi_2[ dX.size( ) * i + I ] += gradChi[ dX.size( ) * dX.size( ) * i + dX.size( ) * I + J ] * dX[ J ];
 
                 }
 
@@ -79,8 +79,10 @@ namespace tractionSeparation{
     }
 
     errorOut computeCurrentDistance( const floatVector &Xi_1, const floatVector &Xi_2, const floatVector &D,
-                                     const floatVector &F,    const floatVector &chi,  const floatVector &grad_chi,
-                                     floatVector &d, floatMatrix &dddF, floatMatrix &dddChi, floatMatrix &dddGradChi ){
+                                     const floatVector &F,    const floatVector &chi,  const floatVector &gradChi,
+                                     floatVector &d,
+                                     floatMatrix &dddXi_1, floatMatrix &dddXi_2, floatMatrix &dddD,
+                                     floatMatrix &dddF, floatMatrix &dddChi, floatMatrix &dddGradChi ){
         /*!
          * Compute the distance in the current configuration where
          * 
@@ -99,8 +101,11 @@ namespace tractionSeparation{
          * \param &D: The initial separation between the particles
          * \param &F: The deformation gradient \f$\frac{dx_i}{dX_I}\f$
          * \param &chi: The micro-deformation
-         * \param &grad_chi: the gradient of the micro deformation w.r.t. the reference spatial variable
+         * \param &gradChi: the gradient of the micro deformation w.r.t. the reference spatial variable
          * \param &d: The current separation between the particles
+         * \param &dddXi_1: The gradient of the separation w.r.t. the local reference micro-position vector
+         * \param &dddXi_2: The gradient of the separation w.r.t. the non-local reference micro-position vector
+         * \param &dddD: The gradient of the separation w.r.t. the local reference separation distance
          * \param &dddF: The gradient of the separation w.r.t. the deformation gradient
          * \param &dddChi: The gradient of the separation w.r.t. the micro deformation
          * \param &dddGradChi: The gradient of the separation w.r.t. the gradient of the micro deformation
@@ -112,8 +117,9 @@ namespace tractionSeparation{
         floatVector chi_2 = chi;
 
         floatMatrix dxdF( dX.size( ), floatVector( F.size( ), 0 ) );
+        floatMatrix dxdX( dX.size( ), floatVector( dX.size( ), 0 ) );
 
-        floatMatrix dchi_2dGradChi( chi.size( ), floatVector( grad_chi.size( ), 0 ) );
+        floatMatrix dchi_2dGradChi( chi.size( ), floatVector( gradChi.size( ), 0 ) );
         floatMatrix dchi_2dX( chi.size( ), floatVector( dX.size( ), 0 ) );
 
         floatVector eye( dX.size( ) * dX.size( ), 0 );
@@ -125,11 +131,15 @@ namespace tractionSeparation{
 
                 dx[ i ] += F[ dX.size( ) * i + I ] * dX[ I ];
 
+                dxdX[ i ][ I ] += F[ dX.size( ) * i + I ];
+
                 for ( unsigned int J = 0; J < dX.size( ); J++ ){
 
-                    chi_2[ dX.size( ) * i + I ] += grad_chi[ dX.size( ) * dX.size( ) * i + dX.size( ) * I + J ] * dX[ J ];
+                    chi_2[ dX.size( ) * i + I ] += gradChi[ dX.size( ) * dX.size( ) * i + dX.size( ) * I + J ] * dX[ J ];
 
                     dxdF[ i ][ dX.size( ) * I + J ] += eye[ dX.size( ) * i + I ] * dX[ J ];
+
+                    dchi_2dX[ dX.size( ) * i + I ][ J ] += gradChi[ dX.size( ) * dX.size( ) * i + dX.size( ) * I + J ];
 
                     for ( unsigned int k = 0; k < dX.size( ); k++ ){
 
@@ -153,12 +163,18 @@ namespace tractionSeparation{
         floatMatrix dxi_1dChi( dX.size( ), floatVector( chi.size( ), 0 ) );
         floatMatrix dxi_2dChi_2( dX.size( ), floatVector( chi.size( ), 0 ) );
 
+        floatMatrix dxi_1dXi_1( dX.size( ), floatVector( Xi_1.size( ), 0 ) );
+        floatMatrix dxi_2dXi_2( dX.size( ), floatVector( Xi_2.size( ), 0 ) );
+
         for ( unsigned int i = 0; i < dX.size( ); i++ ){
 
             for ( unsigned int I = 0; I < dX.size( ); I++ ){
 
                 xi_1[ i ] += chi[ dX.size( ) * i + I ] * Xi_1[ I ];
                 xi_2[ i ] += chi_2[ dX.size( ) * i + I ] * Xi_2[ I ];
+
+                dxi_1dXi_1[ i ][ I ] += chi[ dX.size( ) * i + I ];
+                dxi_2dXi_2[ i ][ I ] += chi_2[ dX.size( ) * i + I ];
 
                 for ( unsigned int K = 0; K < dX.size( ); K++ ){
 
@@ -179,16 +195,209 @@ namespace tractionSeparation{
 
         dddGradChi = vectorTools::dot( dxi_2dChi_2, dchi_2dGradChi );
 
+        dddXi_1 = dxdX - dxi_1dXi_1 + vectorTools::dot( dxi_2dChi_2, dchi_2dX );
+
+        dddXi_2 =  -dxdX + dxi_2dXi_2 - vectorTools::dot( dxi_2dChi_2, dchi_2dX );
+
+        dddD    =  dxdX + vectorTools::dot( dxi_2dChi_2, dchi_2dX );
+
         return NULL;
 
     }
 
     errorOut computeCurrentDistance( const floatVector &Xi_1, const floatVector &Xi_2, const floatVector &D,
-                                     const floatVector &F,    const floatVector &chi,  const floatVector &grad_chi,
-                                     floatVector &d, floatMatrix &dddF, floatMatrix &dddChi, floatMatrix &dddGradChi,
-                                     floatMatrix &d2ddFdF, floatMatrix &d2ddFdChi, floatMatrix &d2ddFdGradChi,
-                                     floatMatrix &d2ddChidChi, floatMatrix &d2ddChidGradChi,
-                                     floatMatrix &d2ddGradChidGradChi );
+                                     const floatVector &F,    const floatVector &chi,  const floatVector &gradChi,
+                                     floatVector &d,
+                                     floatMatrix &dddXi_1, floatMatrix &dddXi_2, floatMatrix &dddD,
+                                     floatMatrix &dddF, floatMatrix &dddChi, floatMatrix &dddGradChi,
+                                     floatMatrix &d2ddFdXi_1,       floatMatrix &d2ddFdXi_2,       floatMatrix &d2ddFdD,
+                                     floatMatrix &d2ddChidXi_1,     floatMatrix &d2ddChidXi_2,     floatMatrix &d2ddChidD,
+                                     floatMatrix &d2ddGradChidXi_1, floatMatrix &d2ddGradChidXi_2, floatMatrix &d2ddGradChidD ){
+        /*!
+         * Compute the distance in the current configuration where
+         * 
+         * \f$d_i = dx_i - \xi_i^1 + \xi_i^2 \f$
+         * 
+         * \f$dx_i = F_{iI} dX_I \f$
+         * 
+         * \f$\xi_i^1 = \chi_{iI} \Xi_I^1 \f$
+         * 
+         * \f$\xi_i^2 = \left(\chi_{iI} + \chi_{iI,J} dX_J \right) \Xi_I^2 \f$
+         * 
+         * \f$dX_I = \Xi_I^1 + D_I - \Xi_I^2 \f$
+         * 
+         * \param &Xi_1: The micro-position vector for the local particle
+         * \param &Xi_2: The micro-position vector for the non-local particle
+         * \param &D: The initial separation between the particles
+         * \param &F: The deformation gradient \f$\frac{dx_i}{dX_I}\f$
+         * \param &chi: The micro-deformation
+         * \param &gradChi: the gradient of the micro deformation w.r.t. the reference spatial variable
+         * \param &d: The current separation between the particles
+         * \param &dddF: The gradient of the separation w.r.t. the deformation gradient
+         * \param &dddChi: The gradient of the separation w.r.t. the micro deformation
+         * \param &dddGradChi: The gradient of the separation w.r.t. the gradient of the micro deformation
+         * \param &d2ddFdXi_1: The second derivative of d w.r.t. F and \f$\Xi_1\f$
+         * \param &d2ddFdXi_2: The second derivative of d w.r.t. F and \f$\Xi_2\f$
+         * \param &d2ddFdD: The second derivative of d w.r.t. F and D
+         * \param &d2ddChidXi_1: The second derivative of d w.r.t. \f$\chi\f$ and \f$\Xi_1\f$
+         * \param &d2ddChidXi_2: The second derivative of d w.r.t. \f$\chi\f$ and \f$\Xi_2\f$
+         * \param &d2ddChidD: The second derivative of d w.r.t. \f$\chi\f$ and D
+         * \param &d2ddGradChidXi_1: The second derivative of d w.r.t. \f$\nabla_X \chi\f$ and \f$\Xi_1\f$
+         * \param &d2ddGradChidXi_2: The second derivative of d w.r.t. \f$\nabla_X \chi\f$ and \f$\Xi_2\f$
+         * \param &d2ddGradChidD: The second derivative of d w.r.t. \f$\nabla_X \chi\f$ and D
+         */
+
+        floatVector dX = Xi_1 + D - Xi_2;
+
+        floatVector dx( dX.size( ), 0 );
+        floatVector chi_2 = chi;
+
+        floatMatrix dxdF( dX.size( ), floatVector( F.size( ), 0 ) );
+        floatMatrix dxdX( dX.size( ), floatVector( dX.size( ), 0 ) );
+
+        floatMatrix d2xdFdX( dX.size( ), floatVector( F.size( ) * dX.size( ), 0 ) );
+
+        floatMatrix dchi_2dGradChi( chi.size( ), floatVector( gradChi.size( ), 0 ) );
+        floatMatrix dchi_2dX( chi.size( ), floatVector( dX.size( ), 0 ) );
+
+        floatMatrix d2chi_2dGradChidX( chi.size( ), floatVector( gradChi.size( ) * dX.size( ), 0 ) );
+
+        floatVector eye( dX.size( ) * dX.size( ), 0 );
+        vectorTools::eye( eye );
+
+        for ( unsigned int i = 0; i < dX.size( ); i++ ){
+
+            for ( unsigned int I = 0; I < dX.size( ); I++ ){
+
+                dx[ i ] += F[ dX.size( ) * i + I ] * dX[ I ];
+
+                dxdX[ i ][ I ] += F[ dX.size( ) * i + I ];
+
+                for ( unsigned int J = 0; J < dX.size( ); J++ ){
+
+                    chi_2[ dX.size( ) * i + I ] += gradChi[ dX.size( ) * dX.size( ) * i + dX.size( ) * I + J ] * dX[ J ];
+
+                    dxdF[ i ][ dX.size( ) * I + J ] += eye[ dX.size( ) * i + I ] * dX[ J ];
+
+                    dchi_2dX[ dX.size( ) * i + I ][ J ] += gradChi[ dX.size( ) * dX.size( ) * i + dX.size( ) * I + J ];
+
+                    for ( unsigned int k = 0; k < dX.size( ); k++ ){
+
+                        d2xdFdX[ i ][ dX.size( ) * dX.size( ) * I + dX.size( ) * J + k ] += eye[ dX.size( ) * i + I ] * eye[ dX.size( ) * J + k ];
+
+                        for ( unsigned int K = 0; K < dX.size( ); K++ ){
+
+                            dchi_2dGradChi[ dX.size( ) * i + I ][ dX.size( ) * dX.size( ) * k + dX.size( ) * K + J ] += eye[ dX.size( ) * i + k ] * eye[ dX.size( ) * I + K ] * dX[ J ];
+
+                            for ( unsigned int L = 0; L < dX.size( ); L++ ){
+
+                                d2chi_2dGradChidX[ dX.size( ) * i + I ][ dX.size( ) * dX.size( ) * dX.size( ) * k + dX.size( ) * dX.size( ) * K + dX.size( ) * J + L ]
+                                    += eye[ dX.size( ) * i + k ] * eye[ dX.size( ) * I + K ] * eye[ dX.size( ) * J + L ];
+
+                            }
+
+                        } 
+
+                    }
+
+                }
+
+            }
+
+        }
+
+        floatVector xi_1( dX.size( ), 0 );
+        floatVector xi_2( dX.size( ), 0 );
+
+        floatMatrix dxi_1dChi( dX.size( ), floatVector( chi.size( ), 0 ) );
+        floatMatrix dxi_2dChi_2( dX.size( ), floatVector( chi.size( ), 0 ) );
+
+        floatMatrix d2xi_1dChidXi_1( dX.size( ), floatVector( chi.size( ) * Xi_1.size( ), 0 ) );
+        floatMatrix d2xi_2dChi_2dXi_2( dX.size( ), floatVector( chi.size( ) * Xi_1.size( ), 0 ) );
+
+        floatMatrix dxi_1dXi_1( dX.size( ), floatVector( Xi_1.size( ), 0 ) );
+        floatMatrix dxi_2dXi_2( dX.size( ), floatVector( Xi_2.size( ), 0 ) );
+
+        for ( unsigned int i = 0; i < dX.size( ); i++ ){
+
+            for ( unsigned int I = 0; I < dX.size( ); I++ ){
+
+                xi_1[ i ] += chi[ dX.size( ) * i + I ] * Xi_1[ I ];
+                xi_2[ i ] += chi_2[ dX.size( ) * i + I ] * Xi_2[ I ];
+
+                dxi_1dXi_1[ i ][ I ] += chi[ dX.size( ) * i + I ];
+                dxi_2dXi_2[ i ][ I ] += chi_2[ dX.size( ) * i + I ];
+
+                for ( unsigned int K = 0; K < dX.size( ); K++ ){
+
+                    dxi_1dChi[ i ][ dX.size( ) * I + K ] += eye[ dX.size( ) * i + I ] * Xi_1[ K ];
+                    dxi_2dChi_2[ i ][ dX.size( ) * I + K ] += eye[ dX.size( ) * i + I ] * Xi_2[ K ];
+
+                    for ( unsigned int L = 0; L < dX.size( ); L++ ){
+
+                        d2xi_1dChidXi_1[ i ][ dX.size( ) * dX.size( ) * I + dX.size( ) * K + L ] += eye[ dX.size( ) * i + I ] * eye[ dX.size( ) * K + L ];
+                        d2xi_2dChi_2dXi_2[ i ][ dX.size( ) * dX.size( ) * I + dX.size( ) * K + L ] += eye[ dX.size( ) * i + I ] * eye[ dX.size( ) * K + L ];
+
+                    }
+
+                }
+
+            }
+
+        }
+
+        d = dx - xi_1 + xi_2;
+
+        dddF = dxdF;
+
+        dddChi = -dxi_1dChi + dxi_2dChi_2;
+
+        dddGradChi = vectorTools::dot( dxi_2dChi_2, dchi_2dGradChi );
+
+        dddXi_1 = dxdX - dxi_1dXi_1 + vectorTools::dot( dxi_2dChi_2, dchi_2dX );
+
+        dddXi_2 =  -dxdX + dxi_2dXi_2 - vectorTools::dot( dxi_2dChi_2, dchi_2dX );
+
+        dddD    =  dxdX + vectorTools::dot( dxi_2dChi_2, dchi_2dX );
+
+        d2ddFdXi_1       = d2xdFdX;
+        d2ddChidXi_1     = -d2xi_1dChidXi_1;
+        d2ddGradChidXi_1 = vectorTools::dot( dxi_2dChi_2, d2chi_2dGradChidX );
+
+        d2ddFdXi_2       = -d2xdFdX;
+        d2ddChidXi_2     = d2xi_2dChi_2dXi_2;
+        d2ddGradChidXi_2 = -vectorTools::dot( dxi_2dChi_2, d2chi_2dGradChidX );
+
+        d2ddFdD          = d2xdFdX;
+        d2ddChidD        = floatMatrix( d.size( ), floatVector( chi.size( ) * D.size( ), 0 ) );
+        d2ddGradChidD    = vectorTools::dot( dxi_2dChi_2, d2chi_2dGradChidX );
+
+        for ( unsigned int i = 0; i < dX.size( ); i++ ){
+            for ( unsigned int j = 0; j < dX.size( ); j++ ){
+                for ( unsigned int J = 0; J < dX.size( ); J++ ){
+                    for ( unsigned int K = 0; K < dX.size( ); K++ ){
+                        for ( unsigned int L = 0; L < dX.size( ); L++ ){
+                            for ( unsigned int a = 0; a < dX.size( ); a++ ){
+                                for ( unsigned int A = 0; A < dX.size( ); A++ ){
+                                     d2ddGradChidXi_2[ i ][ dX.size( ) * dX.size( ) * dX.size( ) * j + dX.size( ) * dX.size( ) * J + dX.size( ) * K + L ]
+                                         += d2xi_2dChi_2dXi_2[ i ][ dX.size( ) * dX.size( ) * a + dX.size( ) * A + L ] * dchi_2dGradChi[ dX.size( ) * a + A ][ dX.size( ) * dX.size( ) * j + dX.size( ) * J + K ];
+                                }
+
+                            }
+
+                        }
+
+                    }
+
+                }
+
+            }
+
+        }
+
+        return NULL;
+
+    }
 
     errorOut decomposeVector( const floatVector &d, const floatVector &n,
                               floatVector &dn, floatVector &dt );