+++ docs/sphinx/changelog.rst
@@ -12,3 +12,4 @@ Changelog
 New Features
 ============
 - Added linear elasticity for the internal particle (:merge:`1`). By `Nathan Miller`_.
+- Added the calculation of the current distance between two points in the particles (:merge:`2`). By `Nathan Miller`_.